pyannote-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pyannote.audio==3.3.2
torch>=2.0.0
python-multipart==0.0.12
